+++ internal/model/tag.go
@@ -4,21 +4,21 @@ type Tag string
 
 type TagWithID struct {
 	tag Tag
-	id  int32
+	id  int64
 }
 
 func (t *Tag) String() string {
 	return string(*t)
 }
 
-func NewTagWithID(id int32, tag Tag) *TagWithID {
+func NewTagWithID(id int64, tag Tag) *TagWithID {
 	return &TagWithID{
 		tag: tag,
 		id:  id,
 	}
 }
 
-func (t *TagWithID) GetID() int32 {
+func (t *TagWithID) GetID() int64 {
 	return t.id
 }
